Fix tests and move web/apps and web/auth tests...
...to their own packages
Also removed the routing package which was introduced also to
avoid circular dependencies in tests and not for a particular
other reasonable argument.
