+++ web/auth_test/auth_test.go
@@ -1,4 +1,7 @@
-package auth
+// auth_test package is introduced to avoid circular dependencies since this
+// particular test requires to depend on routing directly to expose the API
+// and the APP server.
+package auth_test
 
 import (
 	"bytes"
@@ -26,6 +29,7 @@ import (
 	"github.com/cozy/cozy-stack/pkg/sessions"
 	"github.com/cozy/cozy-stack/web"
 	"github.com/cozy/cozy-stack/web/apps"
+	"github.com/cozy/cozy-stack/web/auth"
 	"github.com/cozy/cozy-stack/web/errors"
 	"github.com/cozy/cozy-stack/web/middlewares"
 	"github.com/labstack/echo"
@@ -1029,6 +1033,8 @@ func TestIsLoggedOutAfterLogout(t *testing.T) {
 }
 
 func TestMain(m *testing.M) {
+	config.UseTestFile()
+	config.GetConfig().Assets = "../../assets"
 	instanceURL, _ = url.Parse("https://" + domain + "/")
 	j, _ := cookiejar.New(nil)
 	jar = &testJar{
@@ -1038,7 +1044,6 @@ func TestMain(m *testing.M) {
 		CheckRedirect: noRedirect,
 		Jar:           jar,
 	}
-	config.UseTestFile()
 	instance.Destroy(domain)
 	testInstance, _ = instance.Create(&instance.Options{
 		Domain: domain,
@@ -1055,7 +1060,7 @@ func TestMain(m *testing.M) {
 	r.Renderer = &renderer{
 		t: template.Must(template.ParseGlob("../../assets/templates/*.html")),
 	}
-	Routes(r.Group("/auth", mws...))
+	auth.Routes(r.Group("/auth", mws...))
 
 	r.GET("/test", func(c echo.Context) error {
 		var content string
@@ -1067,7 +1072,7 @@ func TestMain(m *testing.M) {
 		return c.String(http.StatusOK, content)
 	}, mws...)
 
-	handler, err := web.Create(r, apps.Serve)
+	handler, err := web.CreateSubdomainProxy(r, apps.Serve)
 	if err != nil {
 		fmt.Println(err)
 		os.Exit(1)
