+++ web/middlewares/secure_test.go
@@ -1,6 +1,7 @@
 package middlewares
 
 import (
+	"crypto/tls"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -13,13 +14,14 @@ import (
 func TestSecureMiddlewareHSTS(t *testing.T) {
 	e := echo.New()
 	req, _ := http.NewRequest(echo.GET, "http://app.cozy.local/", nil)
+	req.TLS = &tls.ConnectionState{} // fors IsTls to return true
 	rec := httptest.NewRecorder()
 	c := e.NewContext(req, rec)
 	h := Secure(&SecureConfig{
 		HSTSMaxAge: 3600 * time.Second,
 	})(echo.NotFoundHandler)
 	h(c)
-	assert.Equal(t, "max-age=3600; includeSubdomains", rec.Header().Get(echo.HeaderStrictTransportSecurity))
+	assert.Equal(t, "max-age=3600; includeSubDomains", rec.Header().Get(echo.HeaderStrictTransportSecurity))
 }
 
 func TestSecureMiddlewareCSP(t *testing.T) {
